+++ skrub/_table_vectorizer.py
@@ -36,7 +36,7 @@ def _infer_date_format(date_column: pd.Series, n_trials: int = 100) -> str | Non
 
     Parameters
     ----------
-    date_column : :obj:`~pandas.Series`
+    date_column : Series
         A column of dates, as strings.
     n_trials : int, default=100
         Number of rows to use to infer the date format.
@@ -141,7 +141,7 @@ def _replace_missing_in_cat_col(ser: pd.Series, value: str = "missing") -> pd.Se
     replaces the missing values, and returns it.
     """
     ser = _replace_false_missing(ser)
-    if pd.api.types.is_categorical_dtype(ser) and (value not in ser.cat.categories):
+    if isinstance(ser.dtype, pd.CategoricalDtype) and (value not in ser.cat.categories):
         ser = ser.cat.add_categories([value])
     ser = ser.fillna(value=value)
     return ser
@@ -158,9 +158,8 @@ class TableVectorizer(ColumnTransformer):
     Easily transforms a heterogeneous data table
     (such as a :obj:`~pandas.DataFrame`) to a numerical array for machine
     learning. For this it transforms each column depending on its data type.
-    It provides a simplified interface for the
-    :class:`~sklearn.compose.ColumnTransformer`; more documentation of
-    attributes and functions are available in its doc.
+    It provides a simplified interface for the ColumnTransformer ;
+    more documentation of attributes and functions are available in its doc.
 
     .. versionadded:: 0.2.0
 
@@ -179,14 +178,13 @@ class TableVectorizer(ColumnTransformer):
     low_card_cat_transformer : {'drop', 'remainder', 'passthrough'} or Transformer, optional
         Transformer used on categorical/string features with low cardinality
         (threshold is defined by `cardinality_threshold`).
-        Can either be a transformer object instance
-        (e.g. :class:`~sklearn.preprocessing.OneHotEncoder`),
-        a :class:`~sklearn.pipeline.Pipeline` containing the preprocessing steps,
+        Can either be a transformer object instance (e.g. OneHotEncoder),
+        a Pipeline containing the preprocessing steps,
         'drop' for dropping the columns,
         'remainder' for applying `remainder`,
         'passthrough' to return the unencoded columns,
         or `None` to use the default transformer
-        (:class:`~sklearn.preprocessing.OneHotEncoder(handle_unknown="ignore", drop="if_binary")`).
+        (OneHotEncoder(handle_unknown="ignore", drop="if_binary")).
         Features classified under this category are imputed based on the
         strategy defined with `impute_missing`.
 
@@ -194,21 +192,18 @@ class TableVectorizer(ColumnTransformer):
         Transformer used on categorical/string features with high cardinality
         (threshold is defined by `cardinality_threshold`).
         Can either be a transformer object instance
-        (e.g. :class:`~skrub.GapEncoder`),
-        a :class:`~sklearn.pipeline.Pipeline` containing the preprocessing steps,
+        (e.g. GapEncoder), a Pipeline containing the preprocessing steps,
         'drop' for dropping the columns,
         'remainder' for applying `remainder`,
         'passthrough' to return the unencoded columns,
-        or `None` to use the default transformer
-        (:class:`~skrub.GapEncoder(n_components=30)`).
+        or `None` to use the default transformer (GapEncoder(n_components=30)).
         Features classified under this category are imputed based on the
         strategy defined with `impute_missing`.
 
     numerical_transformer : {'drop', 'remainder', 'passthrough'} or Transformer, optional
         Transformer used on numerical features.
-        Can either be a transformer object instance
-        (e.g. :class:`~sklearn.preprocessing.StandardScaler`),
-        a :class:`~sklearn.pipeline.Pipeline` containing the preprocessing steps,
+        Can either be a transformer object instance (e.g. StandardScaler),
+        a Pipeline containing the preprocessing steps,
         'drop' for dropping the columns,
         'remainder' for applying `remainder`,
         'passthrough' to return the unencoded columns,
@@ -218,13 +213,12 @@ class TableVectorizer(ColumnTransformer):
 
     datetime_transformer : {'drop', 'remainder', 'passthrough'} or Transformer, optional
         Transformer used on datetime features.
-        Can either be a transformer object instance
-        (e.g. :class:`~skrub.DatetimeEncoder`),
-        a :class:`~sklearn.pipeline.Pipeline` containing the preprocessing steps,
+        Can either be a transformer object instance (e.g. DatetimeEncoder),
+        a Pipeline containing the preprocessing steps,
         'drop' for dropping the columns,
         'remainder' for applying `remainder`,
         'passthrough' to return the unencoded columns,
-        or `None` to use the default transformer (:class:`~skrub.DatetimeEncoder()`).
+        or `None` to use the default transformer (DatetimeEncoder()).
         Features classified under this category are not imputed at all
         (regardless of `impute_missing`).
 
@@ -258,7 +252,7 @@ class TableVectorizer(ColumnTransformer):
         Note that using this feature requires that the DataFrame columns
         input at :term:`fit` and :term:`transform` have identical order.
 
-    sparse_threshold : float, default=0.3
+    sparse_threshold : float, default=0.0
         If the output of the different transformers contains sparse matrices,
         these will be stacked as a sparse matrix if the overall density is
         lower than this value. Use `sparse_threshold=0` to always return dense.
@@ -268,7 +262,7 @@ class TableVectorizer(ColumnTransformer):
     n_jobs : int, optional
         Number of jobs to run in parallel.
         ``None`` (the default) means 1 unless in a
-        :obj:`joblib.parallel_backend` context.
+        joblib.parallel_backend context.
         ``-1`` means using all processors.
 
     transformer_weights : dict, optional
@@ -294,7 +288,7 @@ class TableVectorizer(ColumnTransformer):
         ``len(transformers_)==len(transformers)+1``, otherwise
         ``len(transformers_)==len(transformers)``.
 
-    columns_ : :obj:`~pandas.Index`
+    columns_ : pandas.Index
         The fitted array's columns. They are applied to the data passed
         to the `transform` method.
 
@@ -308,30 +302,27 @@ class TableVectorizer(ColumnTransformer):
 
     See Also
     --------
-    :class:`skrub.GapEncoder` :
+    GapEncoder :
         Encodes dirty categories (strings) by constructing latent topics with continuous encoding.
-    :class:`skrub.MinHashEncoder` :
+    MinHashEncoder :
         Encode string columns as a numeric array with the minhash method.
-    :class:`skrub.SimilarityEncoder` :
+    SimilarityEncoder :
         Encode string columns as a numeric array with n-gram string similarity.
 
     Notes
     -----
     The column order of the input data is not guaranteed to be the same
-    as the output data (returned by :func:`~TableVectorizer.transform`).
-    This is a due to the way the :class:`~sklearn.compose.ColumnTransformer`
-    works.
+    as the output data (returned by TableVectorizer.transform).
+    This is a due to the way the ColumnTransformer works.
     However, the output column order will always be the same for different
-    calls to :func:`~TableVectorizer.transform` on a same fitted
-    :class:`TableVectorizer` instance.
+    calls to TableVectorizer.transform on a same fitted TableVectorizer instance.
     For example, if input data has columns ['name', 'job', 'year'], then output
     columns might be shuffled, e.g. ['job', 'year', 'name'], but every call
-    to :func:`~TableVectorizer.transform` on this instance will return this
-    order.
+    to TableVectorizer.transform on this instance will return this order.
 
     Examples
     --------
-    Fit a :class:`TableVectorizer` on an example dataset:
+    Fit a TableVectorizer on an example dataset:
 
     >>> from skrub.datasets import fetch_employee_salaries
     >>> ds = fetch_employee_salaries()
@@ -379,7 +370,7 @@ class TableVectorizer(ColumnTransformer):
         impute_missing: Literal["auto", "force", "skip"] = "auto",
         # The next parameters are inherited from ColumnTransformer
         remainder: Literal["drop", "passthrough"] | TransformerMixin = "passthrough",
-        sparse_threshold: float = 0.3,
+        sparse_threshold: float = 0.0,
         n_jobs: int = None,
         transformer_weights=None,
         verbose: bool = False,
@@ -421,7 +412,7 @@ class TableVectorizer(ColumnTransformer):
         We clone the instances to avoid altering them.
         See the clone function docstring.
         Note: typos are not detected here, they are left in and are detected
-        down the line in :func:`~sklearn.compose.ColumnTransformer.fit_transform`.
+        down the line in ColumnTransformer.fit_transform.
         """
         if isinstance(self.low_card_cat_transformer, sklearn.base.TransformerMixin):
             self.low_card_cat_transformer_ = clone(self.low_card_cat_transformer)
@@ -667,16 +658,16 @@ class TableVectorizer(ColumnTransformer):
         for col, dtype in self.types_.items():
             # if categorical, add the new categories to prevent
             # them to be encoded as nan
-            if pd.api.types.is_categorical_dtype(dtype):
+            if isinstance(dtype, pd.CategoricalDtype):
                 known_categories = dtype.categories
                 new_categories = pd.unique(X[col])
+                # remove nan from new_categories
+                new_categories = new_categories[~pd.isnull(new_categories)]
                 dtype = pd.CategoricalDtype(
                     categories=known_categories.union(new_categories)
                 )
                 self.types_[col] = dtype
-            X[col] = X[col].astype(
-                dtype
-            )  # we dont use .loc (https://github.com/pandas-dev/pandas/issues/53729)
+        X = X.astype(self.types_)
         return X
 
     def fit_transform(self, X: ArrayLike, y: ArrayLike = None) -> ArrayLike:
@@ -684,7 +675,7 @@ class TableVectorizer(ColumnTransformer):
 
         In practice, it (1) converts features to their best possible types
         if `auto_cast=True`, (2) classify columns based on their data type,
-        (3) replaces "false missing" (see :func:`_replace_false_missing`),
+        (3) replaces "false missing" (see _replace_false_missing),
         and imputes categorical columns depending on `impute_missing`, and
         finally, transforms `X`.
 
@@ -726,6 +717,17 @@ class TableVectorizer(ColumnTransformer):
         # as we might have some false missing
         # in numerical columns for instance.
         X = _replace_false_missing(X)
+
+        ###
+        # We need to check for duplicate column names.
+        # It is checked by comparing the number of unique values
+        # to the length of the column names
+        if len(set(X.columns)) != len(X.columns):
+            raise AssertionError(
+                "Duplicate column names in the dataframe."
+                f"The column names are {X.columns}"
+            )
+        ###
 
         # If auto_cast is True, we'll find and apply the best possible type
         # to each column.
@@ -864,8 +866,7 @@ class TableVectorizer(ColumnTransformer):
         """Return clean feature names.
 
         Feature names are formatted like:
-        "<column_name>_<value>" if encoded by
-        :class:`~sklearn.preprocessing.OneHotEncoder` or alike,
+        "<column_name>_<value>" if encoded by OneHotEncoder or alike,
         (e.g. "job_title_Police officer"), or "<column_name>" otherwise.
 
         Parameters