+++ CHANGES.rst
@@ -41,6 +41,10 @@ Minor changes
   and :func:`fetch_ken_types` have been renamed.
   :pr:`602` by :user:`Jovan Stojanovic <jovan-stojanovic>`
 
+* Make `pyarrow` an optional dependencies to facilitate the integration
+  with `pyodide`.
+  :pr:`639` by :user:`Guillaume Lemaitre <glemaitre>`.
+
 * Bumped minimal required Python version to 3.10. :pr:`606` by
   :user:`Gael Varoquaux <GaelVaroquaux>`
 
@@ -85,6 +89,13 @@ Minor changes
 * :class:`TableVectorizer` now handles mixed types columns without failing
   by converting them to string before type inference.
   :pr:`623`by :user:`Leo Grinsztajn <LeoGrin>`
+
+* Fixed bug when using :class:`TableVectorizer`'s `transform` method on
+  categorical columns with missing values.
+  :pr:`644` by :user:`Leo Grinsztajn <LeoGrin>`
+
+* :class:`TableVectorizer` never output a sparse matrix by default. This can be changed by
+  increasing the `sparse_threshold` parameter. :pr:`646` by :user:`Leo Grinsztajn <LeoGrin>`
 
 Before skrub: dirty_cat
 ========================